Add ssh_retry connection plugin
The default ssh connection plugin will cause a task to fail if a
connection cannot be made first time. The failures have been found to
cause a number of builds to fail.

This patch adds a new connection plugin called ssh_retry and sets it as
the default one to use. The plugin is a modified version of the ssh
connection plugin from ansible v1.6.10.

This patch removes a retry function in the aio check, the successerator
function was only added to work around the ssh connection failures.

Change-Id: Ic187fb154cfa7b6fa95b19bee4757ec976f3f368
Co-Authored-By: Hugh Saunders <hugh@wherenow.org>
Closes-Bug: #1404343