Enable udev for lvm in cinder-volumes container
There have been reports of cinder performing operations on the wrong
volume. This appears to be due to inconsistencies in the device
numbers recorded for the logical volume. The current lvm.conf disables
the use of udev however the udev service is still running in the
container and on the host. This appears to have created the potential
for errors to creep in under high load.

LVM's use of udev was disabled due to the potential that the host would
be saturated with uevents from containers. The cinder-volumes container
runs on its own host and so this should not be an issue.

This commit adjusts the lvm.conf file created so that udev is used. It
also adds a mount entry to create a devtmpfs on /dev. Finally
'udevadm trigger' is run to add the devices under /dev/mapper.

Closes-Bug: #1436999
Change-Id: I9ab35cf4438a369563f8c08870c1acfd0cc394b0