Serialise rabbitmq playbook to allow upgrades
The rabbitmq playbook is designed to run in parallel across the cluster.
This causes an issue when upgrading rabbitmq to a new major or minor
version because the last node in the cluster that is stopped should be
the first that is started. See
https://www.rabbitmq.com/clustering.html#upgrading for further details.
The current configuration takes no account of this.

This commit serialises the rabbitmq setup play. This means if a new
patch version is installed the whole cluster is not restarted at the
same time. It also allow the last node stopped to be the first that is
started when upgrading to a new major or minor version.

A new variable has been added called rabbitmq_upgrade_major_minor. This
is set to false be default to prevent a new major or minor version being
installed unintentionally. To run the upgrade, which will shutdown the
cluster, the variable can be set to true on the commandline:

  openstack-ansible -e rabbitmq_upgrade_major_minor=true \
  rabbitmq-install.yml

Doc-Impact
Change-Id: I17d4429b9b94d47c1578dd58a2fb20698d1fe02e
Closes-bug: #1474992