WIP: Generic ini conf template
Currently ini-style conf files require explicit templates. This has a
number of downsides in terms of maintenance and configurability.

This commit adds a role to the project for generating a conf file based
on a generic template.

The template requires the ansible variables be named using a specific
format:

prefix__confsection__confvariable: confvariablevalue

For example:

  nova_conf__DEFAULT__verbose: true

__ is the default separator, it is used to split the ansible variable
into the parts for use in the template.

Ansible variables are only allowed to contain [a-zA-Z0-9_]. In most
cases that is sufficient however there a conf files where section names
contain other characters e.g. ':' or '-'. This means there needs to be a
way to allow them to be inserted when required. The role will allow
mappings to be specified for the purpose.

An example of how an option in the swift.conf file would be handled:

  swift_conf__swift__hyphen__hash__swift_hash_path_prefix = "changeme"

The prefix is swift_conf, swift__hyphen__hash is the section in the conf
file, swift_hash_path_prefix is the option name and "chanageme" is the
value.

The config Ansible vars must be found in hostvars[inventory_hostname]
but they may reference vars not found in there.

Implements: blueprint tunable-openstack-configuration
Change-Id: I08e9c4b0e02744fb445691d5a3c2acfd2e83dfc4