Separate Nova services into separate playbooks
Targetting plays at the specific nova groups makes better use of the
functionality of plays by allow the playbook to decide whether or not a
task should be applied to a host. By itself this change does not allow
this to happen, the os_nova role needs to be reworked to remove
conditionals based on group membership.

The ceph_client role is only required by Nova hosts in the nova_compute
group. This commit removes the role from the nova install playbooks that
do not map over nova_compute.
Change-Id: I592c4f8b5eaa265a31ec8c4b85eb8cf21a576e92