Move Nova DB sync into separate playbook
The Nova DB sync performed by nova-manage is part of the os_nova role.
This command only needs to be done on install or when the schema
changes. The OpenStack operations guide also recommends stopping all
Nova services before running the sync.

This commit moves the tasks into a separate playbook so that they can
only run on the required hosts when necessary. The tasks are removed
from the os_nova role because they do not form part of a logical unit
that should be applied to multiple hosts.

A new role, os-nova, has been created so that shared nova variables can
be made available to plays without causing any tasks to be run. It may
be better to instead keep them as group vars but this will work for the
time being.

The new db sync playbook, os-nova-db-sync.yml, is included in
os-nova-install.yml so that the syncronisation is performed on
installation of the nova services.

Change-Id: I36ea6adb6971310b396405647331bee1fb76d196