Stop all but one RabbitMQ node prior to upgrade
RabbitMQ nodes must all be stopped prior to a major/minor version
upgrade [1]. Upgrades can fail when nodes attempt to join the cluster at
the same time. This change uses a play to stop all but one of the nodes,
prior to running the rabbitmq_server role, and then serialises the
running of the role so that one node is upgraded and added to the
cluster at a time. This minimises the downtime as much as possible while
allowing the role to be applied to one node at a time.

[1] http://www.rabbitmq.com/clustering.html#upgrading

Change-Id: I273bc94c7b8195c874ff33a721700c89bd22ff39